+++ evaluation/chunk_jaccard_matrix.py
@@ -1,7 +1,67 @@
-import numpy as np
-from evaluation.helpers import get_intersections_unions, get_confusion_matrix, ConfusionMatrix
-from tensorflow.keras.utils import to_categorical
 from datetime import datetime
+
+import numpy as np
+from tensorflow import keras
+from tensorflow.keras.utils import to_categorical
+
+
+class ConfusionMatrix:
+    __slots__ = [
+        "true_positives",
+        "false_positives",
+        "true_negatives",
+        "false_negatives"
+    ]
+
+    def __init__(self, tp=0, fp=0, tn=0, fn=0):
+        self.true_positives = tp
+        self.false_positives = fp
+        self.true_negatives = tn
+        self.false_negatives = fn
+
+    def add_chunk(self, conf_matrix):
+        self.true_positives += conf_matrix.true_positives
+        self.false_positives += conf_matrix.false_positives
+        self.true_negatives += conf_matrix.true_negatives
+        self.false_negatives += conf_matrix.false_negatives
+
+
+def get_confusion_matrix(y_true, y_pred, label):
+    # revert one hot encoding => binary tensor [0, 0, 1] back to label [2] (3D array to 2D array)
+    label_map_true = np.argmax(y_true, axis=-1)
+    label_map_pred = np.argmax(y_pred, axis=-1)
+    # convert 2D array into 1D array
+    flatten_true = np.reshape(label_map_true, (-1,))
+    flatten_pred = np.reshape(label_map_pred, (-1,))
+
+    tp_mask = (flatten_true == flatten_pred) & (flatten_true == label)
+    true_positives = np.count_nonzero(tp_mask)
+
+    fn_mask = (flatten_true == label) & (flatten_pred != label)
+    false_negatives = np.count_nonzero(fn_mask)
+
+    fp_mask = (flatten_true != label) & (flatten_pred == label)
+    false_positives = np.count_nonzero(fp_mask)
+
+    tn_mask = (flatten_true != label) & (flatten_pred != label)
+    true_negatives = np.count_nonzero(tn_mask)
+
+    return ConfusionMatrix(tp=true_positives, fp=false_positives, tn=true_negatives, fn=false_negatives)
+
+
+def get_intersections_unions(y_true: np.ndarray, y_pred: np.ndarray):
+    intersections = []
+    unions = []
+    for label in range(3):
+        y_true_f = keras.backend.flatten(y_true[..., label])
+        y_pred_f = keras.backend.flatten(y_pred[..., label])
+        intersection = keras.backend.sum(y_true_f * y_pred_f)
+        union = (keras.backend.sum(y_true_f) + keras.backend.sum(y_pred_f) - intersection)
+
+        intersections.append(intersection)
+        unions.append(union)
+
+    return intersections, unions
 
 
 class ChunkJaccardMatrix:
